cleaned up code and fixed login
