+++ app/controllers/users_controller.rb
@@ -1,2 +1,21 @@
 class UsersController < ApplicationController
+    def new
+        @user = User.new
+    end
+
+    def create
+        @user = User.new(user_params)
+        if @user.save
+        session[:user_id] = @user.id
+        redirect_to root_path, notice: "Account created successfully."
+        else
+        render :new
+        end
+    end
+
+    private
+
+    def user_params
+        params.require(:user).permit(:email, :password, :password_confirmation, :first_name, :last_name)
+    end
 end
