+++ app/models/event.rb
@@ -15,22 +15,21 @@
 #  updated_at  :datetime         not null
 #
 class Event < ApplicationRecord
-    has_many :users, through: :guests
-    has_many :guests, through: :guest_lists
-    belongs_to :creator, class_name: "User", foreign_key: "creator_id"
+  has_many :guests, dependent: :destroy
+  belongs_to :creator, class_name: "User", foreign_key: "creator_id"
 
-    validates :title, presence: true
-    validates :start_time, presence: true
-    validates :end_time, presence: true
-    validates :location, presence: true
+  validates :title, presence: true
+  validates :start_time, presence: true
+  validates :end_time, presence: true
+  validates :location, presence: true
 
-    validate :date_must_be_in_the_future
+  validate :date_must_be_in_the_future
 
-    private
+  private
 
-    def date_must_be_in_the_future
-        if date.present? && date < Time.current
-          errors.add(:date, "must be in the future")
-        end
-      end
+  def date_must_be_in_the_future
+    if start_date.present? && start_date < Time.current
+      errors.add(:start_date, "must be in the future")
+    end
+  end
 end
