+++ app/views/home/index.html.erb
@@ -1,4 +1,33 @@
-<h1>Welcome to Eventify</h1>
-<p>
-  <a href="<%= new_user_session_path %>">Login</a>
-</p>
+<h1>Welcome, <%= current_user.first_name %>!</h1>
+<%= button_to "Log Out", logout_path, method: :delete, data: { turbo: false } %>
+
+<% if flash[:notice] %>
+  <div class="alert alert-success"><%= flash[:notice] %></div>
+<% elsif flash[:alert] %>
+  <div class="alert alert-danger"><%= flash[:alert] %></div>
+<% end %>
+
+<%  if @events && @events.any? %>
+  <h2>Here are your events:</h2>
+  <ul>
+    <% @events.each do |event| %>
+      <li>
+        <h3><%= event.title %></h3>
+        <p><strong>Location:</strong> <%= event.location %></p>
+        <p><strong>Description:</strong> <%= event.description %></p>
+        <p><strong>Date:</strong> <%= event.date %></p>
+        <% guest_list_entry = GuestList.find_by(guest_id: current_user.id, event_id: event.id) %>
+        <%= form_with url: update_attendance_guests_path, method: :patch, local: true do |form| %>
+          <%= form.hidden_field :event_id, value: event.id %>
+          <%= form.radio_button :rsvp_status, 'going', checked: guest_list_entry&.rsvp_status == 'going' %>
+          <%= form.label :rsvp_status_going, 'Going' %>
+          <%= form.radio_button :rsvp_status, 'not_going', checked: guest_list_entry&.rsvp_status == 'not_going' %>
+          <%= form.label :rsvp_status_not_going, 'Not Going' %>
+          <%= form.submit 'RSVP' %>
+        <% end %>
+      </li>
+    <% end %>
+  </ul>
+<% else %>
+  <p>You currently have no events.</p>
+<% end %>
