+++ app/views/layouts/application.html.erb
@@ -20,14 +20,6 @@
   </head>
 
   <body>
-    <% if logged_in? %>
-      <p>Welcome, <%= current_user.email %>!</p>
-      <%= link_to "Log Out", logout_path, method: :delete %>
-
-    <% else %>
-      <%= link_to "Log In", user_sessions_path %>
-    <% end %>
-
     <%= yield %>
   </body>
 </html>
