+++ app/views/sessions/new.html.erb
@@ -1,4 +1,5 @@
-<h1>Login</h1>
+<h1>Welcome to Eventify</h1>
+<h3>Login</h1>
 <%= form_with url: user_sessions_path, local: true do |form| %>
   <%= form.label :email %>
   <%= form.email_field :email %>
