+++ app/views/users/new.html.erb
