+++ db/migrate/20241111011600_create_guests.rb
@@ -3,7 +3,7 @@ class CreateGuests < ActiveRecord::Migration[7.2]
     create_table :guests do |t|
       t.string :role, default: "guest"
       t.string :rsvp_status, default: "pending"
-      t.integer :party_size
+      t.integer :party_size, default: 1
       t.references :event, foreign_key: true
       t.references :user, foreign_key: true
