--- db/migrate/20241111014826_create_guest_lists.rb
@@ -1,11 +1,0 @@
-class CreateGuestLists < ActiveRecord::Migration[7.2]
-  def change
-    create_table :guest_lists do |t|
-      t.string :rsvp_status
-      t.references :event, null: false, foreign_key: true
-      t.references :guest, null: false, foreign_key: true
-
-      t.timestamps
-    end
-  end
-end
