--- db/migrate/20241116194915_create_events_guests_join_table.rb
@@ -1,8 +1,0 @@
-class CreateEventsGuestsJoinTable < ActiveRecord::Migration[7.2]
-  def change
-    create_join_table :events, :guests do |t|
-       t.index [ :event_id, :guest_id ]
-       t.index [ :guest_id, :event_id ]
-    end
-  end
-end
